% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_file_ending}
\alias{check_file_ending}
\title{Check the last line of text file}
\usage{
check_file_ending(path, expect)
}
\arguments{
\item{path}{path to the file}

\item{expect}{expected content of the last line}
}
\description{
Check the last line of text file
}
