% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexis.R
\name{check_login}
\alias{check_login}
\title{check login status and move to Power Search}
\usage{
check_login()
}
\description{
This funtion check if the user is logged into Nexis database. If so, move to
Power Search page.
}
