% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{find_element}
\alias{find_element}
\alias{find_elements}
\alias{count_elements}
\title{Select or count elements by XPath}
\usage{
find_element(xpath, wait = FALSE)

find_elements(xpath, wait = FALSE)

count_elements(xpath)
}
\arguments{
\item{xpath}{query to specify elements}

\item{wait}{wait until elements are found if `TRUE`}
}
\description{
Select or count elements by XPath
}
