% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_date_range}
\alias{get_date_range}
\title{Generate date ranges in small chunks}
\usage{
get_date_range(from, to, size = 1, unit = c("year", "month", "week", "day"))
}
\arguments{
\item{from}{first date of the range}

\item{to}{last date of the range}

\item{size}{size of the chunks}

\item{unit}{unit of the size of the chunks}
}
\description{
Generate date ranges in small chunks
}
