% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{get_driver}
\alias{get_driver}
\alias{set_driver}
\title{Get or set Selenium driver (for development)}
\usage{
get_driver()

set_driver(driver)
}
\arguments{
\item{driver}{Selenium driver to use}
}
\description{
Get or set Selenium driver (for development)
}
