% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{get_element_size}
\alias{get_element_size}
\title{Get height and width of element}
\usage{
get_element_size(xpath)
}
\arguments{
\item{xpath}{query to specify element}
}
\description{
Get height and width of element
}
