% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer.R
\name{import_nexis}
\alias{import_nexis}
\title{extract texts and meta data from Nexis HTML files}
\usage{
import_nexis(path, paragraph_separator = "|")
}
\arguments{
\item{path}{either path to a HTML file or a directory that containe HTML files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body texts.}
}
\description{
This extract headings, body texts and meta data (date, byline, length,
secotion, edntion) from items in HTML files downloaded by the scraper.
}
\examples{
irt <- import_nexis('tests/html/irish-times_1995-06-12_0001.html')
afp <- import_nexis('tests/html/afp_2013-03-12_0501.html')
gur <- import_nexis('tests/html/guardian_1986-01-01_0001.html')
all <- import_nexis('tests/html')
}
