% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{is_completed}
\alias{is_completed}
\title{Check if files exists in the data folder}
\usage{
is_completed(prefix, date, ext)
}
\arguments{
\item{prefix}{prefix of the files}

\item{date}{date of the files}

\item{ext}{extension of the files}
}
\description{
Check if files exists in the data folder
}
