% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{make_filename}
\alias{make_filename}
\title{Generate file names}
\usage{
make_filename(prefix, date, range, last, ext)
}
\arguments{
\item{prefix}{prefix of the files}

\item{date}{date of the files}

\item{range}{number indicating the items download}

\item{last}{number indicating the last item}

\item{ext}{extension of the files}
}
\description{
Generate file names
}
