% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{open_browser}
\alias{open_browser}
\alias{navigate_browser}
\title{Open page in browser}
\usage{
open_browser(url, browser = "firefox")

navigate_browser(url)
}
\arguments{
\item{url}{URL to load in the browser}

\item{browser}{web browser to start}
}
\description{
Open page in browser
}
