% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{save_file}
\alias{save_file}
\title{Save downloaded temporary file in a new location}
\usage{
save_file(path)
}
\arguments{
\item{path}{path to a new location}
}
\description{
Save downloaded temporary file in a new location
}
