% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{select_window}
\alias{select_window}
\title{Select browser's pop-up window by window titles}
\usage{
select_window(title)
}
\arguments{
\item{title}{title of the window}
}
\description{
Select browser's pop-up window by window titles
}
