% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_directory}
\alias{set_directory}
\alias{get_directory}
\title{Set or get path to directory to save files}
\usage{
set_directory(path)

get_directory()
}
\arguments{
\item{path}{path to the directory}
}
\description{
Set or get path to directory to save files
}
