% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{submit}
\alias{submit}
\title{set and submit search query}
\usage{
submit(query, date, date_format = "\%m/\%d/\%Y")
}
\arguments{
\item{query}{a query string for Nexis database}

\item{date}{a vector Date object created by \code{get_date_range()}. The fist
date is the first day and second element is the last day of the serach
period.}

\item{format}{format the date. It should be "\%d\%m/\%Y" for Nexis UK, but
"\%m/\%d/\%Y" for other versions.}
}
\description{
This function sets search keyword and data, and submit to the database. Users
should select sources manually before running this command.
}
