% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{wait_for}
\alias{wait_for}
\title{Wait for element to be found}
\usage{
wait_for(xpath, timeout = 120)
}
\arguments{
\item{xpath}{query to specify element}

\item{timeout}{maximum time to wait in seconds}
}
\description{
Wait for element to be found
}
